+++ ginit/__main__.py
@@ -8,11 +8,10 @@
 #  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 #  copies of the Software, and to permit persons to whom the Software is
 #  furnished to do so.
-
 import ginit
-import ginit.visitor as visitor
 import ginit.generation as generation
 import ginit.patching as patching
+import ginit.visitor as visitor
 
 
 def main():
@@ -33,36 +32,25 @@ def main():
     parser.add_argument('-p', '--patch', action='store_true',
                         help='Patch imports of python files at path',
                         default=False)
-
-    parser.add_argument('--verbose', nargs='?', default=0, type=int,
-                        help='Verbosity level')
-
-    parser.add_argument('-v', '--version', action='store_true',
-                        help='print version and exit')
-
     args, unknown = parser.parse_known_args()
-
-    if args.verbose == 0:
-        level = logging.WARNING
-    elif args.verbose == 1:
-        level = logging.INFO
-    else:
-        level = logging.DEBUG
 
     logging.basicConfig(
         format='%(levelname)s: %(message)s',
-        level=level,
+        level=logging.INFO,
     )
-
-    visit_result = visitor.visit_path(args.path,
-                                      visit_cython=args.cython)
-    if args.patch:
-        patching.patch_imports_form(module_visitor_dict=visit_result,
-                                    is_cython_path=args.cython)
+    
+    if args.patch and args.cython:
+        patching.imports_regex_patcher(args.path, is_cython_path=args.cython)
     else:
-        generation.gen_python_init_from_visit(module_visitor_dict=visit_result,
-                                              in_place=args.inplace,
-                                              is_cython_init=args.cython)
+        visit_result = visitor.visit_path(args.path,
+                                          visit_cython=args.cython)
+        if args.patch:
+            patching.patch_imports_form(module_visitor_dict=visit_result,
+                                        is_cython_path=args.cython)
+        else:
+            generation.gen_python_init_from_visit(module_visitor_dict=visit_result,
+                                                  in_place=args.inplace,
+                                                  is_cython_init=args.cython)
 
 
 if __name__ == '__main__':