+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2020-09-29
+### Added
+- Cython patching support
+
+### Fixed
+- Cython __init__.pxd generation
+
 ## [1.0.0] - 2020-09-28
 ### Added
 - Project basis